Revert "added static map on load and dymamic map will replace it on click (#492)"
This reverts commit 61d02381a1cdc9ef4951e7d6ad4b2112db7833f9.
