+++ package.json
@@ -38,7 +38,6 @@
   },
   "devDependencies": {
     "@babel/cli": "^7.8.4",
-    "@babel/plugin-transform-runtime": "^7.13.15",
     "@babel/preset-env": "^7.9.0",
     "@babel/preset-react": "^7.10.4",
     "babel-loader": "^8.0.5",
