+++ src/components/OutbreakMap/_map.scss
@@ -8,13 +8,6 @@ $map-color-background: #cad2d3;
   height: 500px;
   background: $map-color-background;
   border-radius: $box-border-radius;
-
-  img {
-    width: 100%;
-    height: 100%;
-    border-radius: 10px;
-    cursor: zoom-in;
-  }
 }
 
 #map-legend {
