+++ src/components/OutbreakMap/index.js
@@ -1,9 +1,7 @@
 import drawMap from "./DrawMap";
 import drawMapPrefectures from "./DrawMapPrefectures";
-import { drawLegend } from "./drawLegend";
 
 export default {
   drawMap,
   drawMapPrefectures,
-  drawLegend,
 };
