+++ src/index.html
@@ -139,7 +139,7 @@
       </div>
 
       <div class="kpi-box" id="kpi-deceased">
-        <a href="#deceased" class="inline">
+        <a href="#deaths" class="inline">
           <div class="label" data-i18n="kpi-deceased">Deaths</div>
         </a>
         <div class="value"><div class="lds-dual-ring"></div></div>
@@ -258,7 +258,7 @@
 
     <section class="national-charts embed-hide">
       <section class="daily-increase-container embed-hide">
-        <a name="deceased">
+        <a name="deaths">
           <h4 data-i18n="deceased-chart-caption">Daily Deaths</h4>
         </a>          
         <div class="daily-increase-chart-box">
